Allow users with display permissions on promos to see the edit page.
This allows users with the promotion display permission set to see the
edit page of a promotion for further details of the promtion, as it was
a desired feature for some people. This comes hand in hand with the
ability to download the promotion codes which were also added as
permissions.