Fix displaying out of stock variants for returns
Currently, when creating an exchange, a variant will be displayed as
an eligible exchange variant if there is stock for that variant in
any stock location. This can lead to errors when creating the return
shipment if there isn't any stock in the active stock locations.

This changes adds an optional parameter to the in_stock scope that
determines whether a variant is in stock for the given stock locations.
It also changes the UI for return authorizations and reimbursements to
only display the eligible exchange variants that have stock in active
stock locations.