Introduce variant level properties
Takes the concept of product properties and transfers it over to
variants.

This change allows an admin to specify properties for variants based on
a selection of option values. This prevents a lot of duplication since
many variants are going to share the same properties.

Unlike product properties which are associated directly to the product,
the variant properties are not directly associated with their variants.
A variant property rule is created which contains the option values the
rule is targeting.

The reason for this choice was due to performance. With products that
have thousands of variants, associating information for each variant
becomes a performance concern. An added benefit to this approach is that
variant properties do not have to be set up for new variant. If the
variant has the option values targeted by the rule, the properties will
automatically apply to the variant.

The UI displays the variant properties on the same page as the product
properties. The JavaScript changes were related to fixing some issues
related to adding new fields to the property list UI. The changes were
related to supporting property lists on the same page.