Associate shipments and payments before progressing the order’s state
Removed the state update that was occurring in set_shipment_for_new_order causing the order to be in the “confirm” state without going through the state machine. I also moved the payment creation up so that the new order has the payment associated with it while transitioning states.

The update_attributes in set_shipment_for_new_order was what caused the UnreturnedItemCharger spec to not fail when the check for an unreturned_exchange isn’t present in order.rb’s create_proposed_shipments.