Remove method_missing from Spree::Gatway
This method missing made it troublesome to develop anything on a
gateway. It also didn't bother to implement respond_to_missing.

This should be redundent with the delegate at the top of the file, which
delegates the standard ActiveMerchant methods to the provider.

These are probably the only methods which were used in this way, and
overriding them in subclasses of gateway should still have the same
behaviour.