Replace State=Active/Inactive with Active=Yes/No
On the generic "Named Types" include, replace the column title of
"State" and values of "Active"/"Inactive" with column title "Active"
and values "Yes"/"No".

This change helps prevent issues with :state that may be localized
to mean a state within a country (e.g. US states). It also standardizes
the look and feel with other similar tables that already use the
model with meaning in the column title and "Yes"/"No" in the values
(such as the "Tax Categories", "Zones" or "Countries" tables).

Only the named_types index had to be modified. The rest of the actions
already use the Active=Yes/No logic which requires no modifications.