improve assigning addresses to unsaved user
This was originally noticed in tests that wanted to create a user
with a ship_address at once. I don't know how common it might happen
in production, but seems a reasonable expectation. The core issue
was the reset at the end of mark_default -- the unsaved UserAddress
was getting dropped. That reset was originally put in place to ensure
ordering of the default address first and to let us re-add a previously
removed address. So additional tests were added and a new approach
found.