Fully update the order after payments are saved.
While we keep the intention of only updating the order when the order or
payment is in a certain state, we shouldn't be doing this manually.

The order updater exposes a method that's responsible for the total
update of an order, a payment shouldn't know that shipments need to be
updated when the payment is saved.

Additionally, this bypasses the order updater hooks, which run as part
of the update process, preventing users who rely on these hooks for
information about when state changes may occur from using the hooks to
accomplish that goal.