FIx a stale data issue when using Spree::OrderCapturing.
As described in the comment, calling `payment.capture!` triggers an
after_save on Spree::Payment. In this after_save, things can happen, one
of those things is an `order.update!`.

That means if the `order.update!` happens within a payment after save, the order we have stored in `@order` is now out of date, and has to be reloaded.

If we don't reload it, the hooks in the order updater will be called
twice. If they make decisions around fresh state changes, idempotency
could be compromised.