Restrict who has the ability to update a user's email address.
There is a possibility for privelege escalation whereby a lower-level
admin changes the email address of a higher-level admin to an email
address they have access to, thereby gaining access to more admin
permissions. To prevent this vulnerability, this code restricts the
ability for admins with roles below SuperUser only to update the email
address for themselves and for users with no roles.

* Refactor permitted_user_attributes, remove email, add nested address attributes that were duplicated
* Permit email attribute on create and when the user has the ability
* Restrict UserManagement permission to not update email for users with roles
* Disable the admin user email field when not authorized to update it