Clean up authorization test helpers.
The current way that authorization is stubbed on feature
specs only allows non-superuser configuration through anonymous blocks.
The addition of permission sets gives us a type of user that we
want to assert against, but we don't yet have a great way of testing the
abilities of that type of user at the controller / feature level, plus
our anonymous authorization changes might not be a great representation
of the behavior we are trying to test.

This change simplifies the authorization helper code,
and allows for the testing of permission sets at a controller/feature
level more easily, while maintaining anonymous abilities at the
controller level where that level of granularity may be more useful.