Allow stock transfer permissions to view stock transfers to/from from viewable stock locations.
When given the stock transfer display or stock transfer management
permissions, there is a dropdown in admin/stock locations to filter by
stock locations, which was filled by the ability to transfer_to or
transfer_from a stock location. This change allows people with the stock
transfer permission sets to use that feature of stock transfers without
having to also have the ability to transfer_to or transfer_from a stock
location.