Allow item_adjustments form to specify calculator-specific fields
Given a promotion action with complex fields, when creating it with and order-level
action it will allow you to specify the template that you want to render for the
promotion action. If no template is available, it uses the default fields.

If the promotion action is for the item-level, it will always use the default fields.

This allows stores to have the same logic for both order and line-item level promotion actions.