Add specs for international delivery
Adds specs for the refund that applies once you ship outside
a VAT jurisdiction. This refund is necessary to arrive at the correct
line item totals as long as we are not able to calculate correct
line item prices.

Also adds support for the situation that the default tax zone is contained by
a rates tax zone, and you're shipping within that. This support is
particularly ugly, as it comes with a very expensive call to
`Spree::Zone.contains?`.