Avoid extra stock location in order factory
Previously, create(:order_with_line_items) would create two stock
locations. One would be created as part of the product factory, which
creates a stock location if none exists, and another would be created as
part of the shipment factory.

This commit moves the creation of the stock location up a level, into
the order factory. This allows it to be created before the line_item and
product factory.

This also allows doing:

    create(:shipped_order, stock_location: location)

which should be helpful in testing.