Fix circleci caching
In the previous commit, I messed up. I needed to specify the path for
the top level bundle install to match that used by the build-ci script.

In my initial testing, I verified that this worked when the bundle
install was done from the build-ci.rb script, and didn't properly verify
when I moved that out to circle.yml.

Oops.