Replace `TaxRate.adjust` with {Item,Order}Adjuster.adjust!
The code in `Spree::TaxRate.adjust!` is hard to understand, and the
adjusting should happen in dedicated adjuster classes. This commit
ports the functionality of `TaxRate.match` over to two classes:

* An `OrderAdjuster` which iterates over all items (line items and shipments)
  in an order and runs the

* `ItemAdjuster`, which actually creates adjustments for every item.

The code should read a lot easier now.

In order to reduce database load, the tax rate array for an order can be passed
into the `ItemAdjuster.new`. Considering the performance benefits, I think
that's justifiable amount of shared knowledge between the two Adjuster classes.