Add a carrier and service level column to Spree::ShippingMethods.
When we're dealing with a shipping method, we have no formal way of
adding additional information that would be useful for fulfillment
purposes.

We can add an "internal" name to the shipping method, but that would
require pattern matching to get any useful information from. Often, in
most fulfillment scenarios, we have a desire to know _exactly_ what
carrier and service level a specific shipping method is representative
of.

Given a shipment I can then infer that it should be shipped with UPS one
day shipping by introspecting on the methods' carrier and service_level
columns as necessary.

This is optional as it stands, but validations can be added in stores as
required for complex fulfillment scenarios.