Extract tax adjuster helpers into module
The order adjuster and the item adjuster share code, because I need to run
some methods on the order level and cache them. This extracts these methods
into a module.

Inheritance would've also worked, but in this case I find a mixin
more fitting as the behaviour of the two classes is still quite different.