Don't automatically add javascripts to dummy app
LIB_NAME is used in the install generator when we are creating the
spec/dummy app in an extension. Previously it added includes for
files like `spree/backend/$LIB_NAME.js` whether or not they existed.

Unfortunately, as of sprockets-rails 3.0, that will cause an exception
even if that section (backend/frontend) of solidus isn't visited.

If that file is needed for the extension to run, it should be added by
the extension's own install generator. Most extensions already do this
and won't be broken by this change.