Use <label class="required"> instead of extra span
There are a few advantages to this.

Previously there was a mix of <span class="required"> inside and outside
of the label. It's now consistent.

When the span and text was inside of the label, capybara would consider
it part of the text of the label. Now it is considered presentation and
not part of the label's text. This required changing one spec which had
fill_in "NAME *".

When the span and text were outside of the label, clicking the span did
not focus the input field. It now does, because the label is the parent
element and the click will bubble up.

How we display a required attribute is a style concern. If we change our
mind about the way a required attribute's label should work we can now
change it everywhere from our CSS.

This is also much shorter and easier to type in a view.