Remove default translation from tab helper
If we do not have a defined translation this should show an missing
translation.

Previously, the tab helper was sometimes passed a key to translate, and
other times a translation from Spree.t. In the latter case, it would
then translate that already translated value, which will of course
behave differently in different languages.

We should have the tab helper always use a translation from
spree.admin.tab.*