removes :force attribute from params
This commit addresses #652 where a user was not able to update the
inventory in the admin panel due to #permitted_stock_item_attributes not
including :force as a valid parameter.

I added the attribute `:backorderable` which allowed our test suite to
pass, however I have concerns that I may be missing something here and
was contemplating whether it would be best to add the attribute in core
or strip out of the param after checking it. I went with stripping it
out of the params.

You can find the reference to attributes here
https://github.com/solidusio/solidus/blob/e15ca73e8f7083633343f0678023695a96e24d31/core/lib/spree/permitted_attributes.rb#L94