Reduce the tabs component's reliance on class names
We can safely assume that all `<li>` elements under the `.tabs`
namespace should be tabs. Therefor we don't need to use a `.tab` class.
This will allow other sidebars to be converted much easier because all
we will need to do is add a single class name to the list.