Refactor Spree::ShippingRate#display_price
This method used to be a monster. With the new ShippingRateTax Model,
we can now add support for multiple taxes on shipping rates.

The calculation of shipping rate tax amounts now goes through
Spree::TaxRate#compute_amount. This way, we don't have to duplicate
the multiplying by (-1) in the tax calculator and further align how
shipping rate taxes are calculated.

This required a serious spec-rewrite.